+++ examples/split_placement/main_ppo_split.py
@@ -132,10 +132,10 @@ def main_task(config):
 
     elif config.actor_rollout_ref.actor.strategy == "megatron":
         assert config.actor_rollout_ref.actor.strategy == config.critic.strategy
-        from verl.single_controller.ray.megatron import NVMegatronRayWorkerGroup
+        from verl.single_controller.ray import RayWorkerGroup
         from verl.workers.megatron_workers import ActorRolloutRefWorker, CriticWorker
 
-        ray_worker_group_cls = NVMegatronRayWorkerGroup
+        ray_worker_group_cls = RayWorkerGroup
 
     else:
         raise NotImplementedError
