+++ recipe/prime/main_prime.py
@@ -81,10 +81,10 @@ def main_task(config, compute_score=None):
 
     elif config.actor_rollout_ref.actor.strategy == "megatron":
         assert config.actor_rollout_ref.actor.strategy == config.critic.strategy
-        from verl.single_controller.ray.megatron import NVMegatronRayWorkerGroup
+        from verl.single_controller.ray import RayWorkerGroup
         from verl.workers.megatron_workers import ActorRolloutRefWorker
 
-        ray_worker_group_cls = NVMegatronRayWorkerGroup
+        ray_worker_group_cls = RayWorkerGroup
 
     else:
         raise NotImplementedError
