+++ tests/single_controller/detached_worker/client.py
@@ -21,8 +21,7 @@ from server import Trainer
 from tensordict import TensorDict
 
 from verl import DataProto
-from verl.single_controller.ray import RayClassWithInitArgs
-from verl.single_controller.ray.megatron import NVMegatronRayWorkerGroup
+from verl.single_controller.ray import RayClassWithInitArgs, RayWorkerGroup
 
 
 def compute_position_id_with_mask(mask):
@@ -34,9 +33,7 @@ if __name__ == "__main__":
     # get the worker group using names
     worker_names = ["trainerTrainer_0:0", "trainerTrainer_0:1"]
     cls_with_init_args = RayClassWithInitArgs(cls=Trainer)
-    worker_group = NVMegatronRayWorkerGroup.from_detached(
-        worker_names=worker_names, ray_cls_with_init=cls_with_init_args
-    )
+    worker_group = RayWorkerGroup.from_detached(worker_names=worker_names, ray_cls_with_init=cls_with_init_args)
 
     batch_size = 16
     sequence_length = 1024
