+++ verl/single_controller/base/decorator.py
@@ -39,11 +39,6 @@ def init_predefined_dispatch_mode():
     Dispatch.register("RANK_ZERO")
     Dispatch.register("ONE_TO_ALL")
     Dispatch.register("ALL_TO_ALL")
-    Dispatch.register("MEGATRON_COMPUTE")
-    Dispatch.register("MEGATRON_PP_AS_DP")
-    Dispatch.register("MEGATRON_PP_ONLY")
-    Dispatch.register("MEGATRON_COMPUTE_PROTO")
-    Dispatch.register("MEGATRON_PP_AS_DP_PROTO")
     Dispatch.register("DP_COMPUTE")
     Dispatch.register("DP_COMPUTE_PROTO")
     Dispatch.register("DP_COMPUTE_PROTO_WITH_FUNC")
@@ -136,64 +131,6 @@ def collect_all_to_all(worker_group, output):
     return output
 
 
-def dispatch_megatron_compute(worker_group, *args, **kwargs):
-    """
-    User passes in dp data. The data is dispatched to all tp/pp ranks with the same dp
-    """
-    from verl.single_controller.base.megatron.worker_group import MegatronWorkerGroup
-
-    assert isinstance(worker_group, MegatronWorkerGroup), (
-        f"worker_group must be MegatronWorkerGroup, Got {type(worker_group)}"
-    )
-
-    # ray put all the args in advance to avoid duplicate serialization cost
-    import ray
-
-    args = [[ray.put(dp_arg) for dp_arg in arg] for arg in args]
-    kwargs = {k: [ray.put(dp_v) for dp_v in v] for k, v in kwargs.items()}
-
-    def _transform_data(obj_list, worker_group):
-        assert isinstance(obj_list, tuple | list) and len(obj_list) == worker_group.dp_size
-        transformed_data = []
-        for i in range(worker_group.world_size):
-            local_dp_rank = worker_group.get_megatron_rank_info(rank=i).dp_rank
-            transformed_data.append(obj_list[local_dp_rank])
-        return transformed_data
-
-    all_args = tuple([_transform_data(arg, worker_group) for arg in args])
-    all_kwargs = {key: _transform_data(val, worker_group) for key, val in kwargs.items()}
-
-    return all_args, all_kwargs
-
-
-def collect_megatron_compute(worker_group, output):
-    """
-    Only collect the data from the tp=0 and pp=last and every dp ranks
-    """
-    from verl.single_controller.base.megatron.worker_group import MegatronWorkerGroup
-
-    assert isinstance(worker_group, MegatronWorkerGroup)
-    output_in_dp = []
-    pp_size = worker_group.get_megatron_global_info().pp_size
-    for global_rank in range(worker_group.world_size):
-        local_rank_info = worker_group.get_megatron_rank_info(rank=global_rank)
-        if local_rank_info.tp_rank == 0 and local_rank_info.pp_rank == pp_size - 1 and local_rank_info.cp_rank == 0:
-            output_in_dp.append(output[global_rank])
-    return output_in_dp
-
-
-def dispatch_megatron_compute_data_proto(worker_group, *args, **kwargs):
-    """
-    All the args and kwargs must be DataProto. The batch will be chunked by dp_size and passed to each rank
-    """
-    from verl.single_controller.base.megatron.worker_group import MegatronWorkerGroup
-
-    assert isinstance(worker_group, MegatronWorkerGroup)
-
-    splitted_args, splitted_kwargs = _split_args_kwargs_data_proto(worker_group.dp_size, *args, **kwargs)
-    return dispatch_megatron_compute(worker_group, *splitted_args, **splitted_kwargs)
-
-
 def _concat_data_proto_or_future(output: list):
     import ray
 
@@ -211,112 +148,6 @@ def _concat_data_proto_or_future(output: list):
         return DataProtoFuture.concat(output)
     else:
         raise NotImplementedError
-
-
-def collect_megatron_compute_data_proto(worker_group, output):
-    """
-    Each output must be a DataProto. We concat the dim=0 of output
-    """
-    import ray
-
-    from verl.protocol import DataProto
-
-    output = collect_megatron_compute(worker_group, output)
-    for o in output:
-        assert isinstance(o, DataProto | ray.ObjectRef), f"expecting {o} to be DataProto, but got {type(o)}"
-
-    return _concat_data_proto_or_future(output)
-
-
-def dispatch_megatron_pp_as_dp(worker_group, *args, **kwargs):
-    """
-    treat pp as dp.
-    """
-    from verl.single_controller.base.megatron.worker_group import MegatronWorkerGroup
-
-    assert isinstance(worker_group, MegatronWorkerGroup)
-
-    pp_size = worker_group.pp_size
-    dp_size = worker_group.dp_size
-    cp_size = worker_group.cp_size
-    pp_dp_cp_size = pp_size * dp_size * cp_size
-
-    def _transform_data(obj_list, worker_group):
-        assert isinstance(obj_list, list | tuple) and len(obj_list) == pp_dp_cp_size
-        transformed_data = []
-        for i in range(worker_group.world_size):
-            local_dp_rank = worker_group.get_megatron_rank_info(rank=i).dp_rank
-            local_pp_rank = worker_group.get_megatron_rank_info(rank=i).pp_rank
-            local_cp_rank = worker_group.get_megatron_rank_info(rank=i).cp_rank
-            # compute the rank in obj_list. Note that the order is dp then cp then pp
-            # Also note that the outputs within a pp group will be firstly allgathered, then only the
-            # output of pp0 will be collected.
-            # For pp=2 dp=4, a batch of data "ABCDEFGH" should be dispatched and collected in below order:
-            #    dispatch:       pp_allgther:        collect:
-            #   dp 0 1 2 3      dp  0  1  2  3
-            # pp +---------+  pp +-------------+
-            #  0 | A C E G |   0 | AB CD EF GH |     ABCDEFGH
-            #  1 | B D F H |   1 | AB CD EF GH |
-            #    +---------+     +-------------+
-            dp_cp_rank = local_cp_rank * dp_size + local_dp_rank
-            arg_rank = dp_cp_rank * pp_size + local_pp_rank
-            transformed_data.append(obj_list[arg_rank])
-        return transformed_data
-
-    all_args = tuple([_transform_data(arg, worker_group) for arg in args])
-    all_kwargs = {key: _transform_data(val, worker_group) for key, val in kwargs.items()}
-
-    return all_args, all_kwargs
-
-
-def collect_megatron_pp_as_dp(worker_group, output):
-    """
-    treat pp as dp. Only collect data on tp=0
-    """
-    from verl.single_controller.base.megatron.worker_group import MegatronWorkerGroup
-
-    assert isinstance(worker_group, MegatronWorkerGroup)
-    output_in_dp = []
-    for global_rank in range(worker_group.world_size):
-        local_rank_info = worker_group.get_megatron_rank_info(rank=global_rank)
-        if local_rank_info.tp_rank == 0:
-            output_in_dp.append(output[global_rank])
-    return output_in_dp
-
-
-def collect_megatron_pp_only(worker_group, output):
-    """
-    Only collect output of megatron pp. This is useful when examine weight names as they are identical in tp/dp
-    """
-    from verl.single_controller.base.megatron.worker_group import MegatronWorkerGroup
-
-    assert isinstance(worker_group, MegatronWorkerGroup)
-    output_in_pp = []
-    for global_rank in range(worker_group.world_size):
-        local_rank_info = worker_group.get_megatron_rank_info(rank=global_rank)
-        if local_rank_info.tp_rank == 0 and local_rank_info.dp_rank == 0:
-            output_in_pp.append(output[global_rank])
-    return output_in_pp
-
-
-def dispatch_megatron_pp_as_dp_data_proto(worker_group, *args, **kwargs):
-    from verl.single_controller.base.megatron.worker_group import MegatronWorkerGroup
-
-    assert isinstance(worker_group, MegatronWorkerGroup)
-
-    pp_dp_cp_size = worker_group.dp_size * worker_group.pp_size * worker_group.cp_size
-    splitted_args, splitted_kwargs = _split_args_kwargs_data_proto(pp_dp_cp_size, *args, **kwargs)
-    ret = dispatch_megatron_pp_as_dp(worker_group, *splitted_args, **splitted_kwargs)
-    return ret
-
-
-def collect_megatron_pp_as_dp_data_proto(worker_group, output):
-    from verl.single_controller.base.megatron.worker_group import MegatronWorkerGroup
-
-    assert isinstance(worker_group, MegatronWorkerGroup)
-
-    output = collect_megatron_pp_as_dp(worker_group, output)
-    return _concat_data_proto_or_future(output)
 
 
 def dispatch_dp_compute(worker_group, *args, **kwargs):
@@ -485,23 +316,6 @@ DISPATCH_MODE_FN_REGISTRY = {
     Dispatch.ALL_TO_ALL: {
         "dispatch_fn": dispatch_all_to_all,
         "collect_fn": collect_all_to_all,
-    },
-    Dispatch.MEGATRON_COMPUTE: {
-        "dispatch_fn": dispatch_megatron_compute,
-        "collect_fn": collect_megatron_compute,
-    },
-    Dispatch.MEGATRON_PP_AS_DP: {
-        "dispatch_fn": dispatch_megatron_pp_as_dp,
-        "collect_fn": collect_megatron_pp_as_dp,
-    },
-    Dispatch.MEGATRON_PP_ONLY: {"dispatch_fn": dispatch_one_to_all, "collect_fn": collect_megatron_pp_only},
-    Dispatch.MEGATRON_COMPUTE_PROTO: {
-        "dispatch_fn": dispatch_megatron_compute_data_proto,
-        "collect_fn": collect_megatron_compute_data_proto,
-    },
-    Dispatch.MEGATRON_PP_AS_DP_PROTO: {
-        "dispatch_fn": dispatch_megatron_pp_as_dp_data_proto,
-        "collect_fn": collect_megatron_pp_as_dp_data_proto,
     },
     Dispatch.DP_COMPUTE: {"dispatch_fn": dispatch_dp_compute, "collect_fn": collect_dp_compute},
     Dispatch.DP_COMPUTE_PROTO: {
